# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisResponse": {
        "additionalProperties": false,
        "description": "The last ``analysis/{instrument}`` snapshot — or an honest empty state.\n\n``state = \"empty\"`` means NO dossier was ever published for this\ninstrument: every snapshot-derived field is ``None`` (never invented)\nand ``reason`` says why. ``state = \"ok\"`` relays the persisted content\nverbatim:\n\n- ``advice`` is the canonical ``AdviceResult`` produced by THE single\n  ``AdviceEngine`` — status, direction, the ten gates with their reason\n  codes, limitations — exactly as published; the API neither recomputes\n  nor softens it;\n- ``bars`` carries the validated synthetic OHLCV series (decimal\n  strings) with its per-bar discard account;\n- ``evidence`` is the fusion-cluster rail of the instrument;\n- ``scenarios`` is either the ``THEORETICAL`` scenario grid with its\n  ``CalculationRecord`` lineage or an honest ``ABSENT`` block with its\n  typed reason.",
        "properties": {
          "advice": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advice"
          },
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "bars": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bars"
          },
          "coverage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "engine_version": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "instrument": {
            "minLength": 1,
            "title": "Instrument",
            "type": "string"
          },
          "population": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "reason": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "scenarios": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenarios"
          },
          "snapshot_version": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Version"
          },
          "state": {
            "enum": [
              "ok",
              "empty"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "snapshot_version",
          "as_of",
          "population",
          "instrument",
          "engine_version",
          "bars",
          "evidence",
          "scenarios",
          "advice",
          "coverage",
          "reason"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionChainResponse": {
        "additionalProperties": false,
        "description": "The last ``option_chain/{underlying}`` snapshot — or an honest empty.\n\n``state = \"empty\"`` means NO snapshot was ever published for this\nunderlying: every snapshot-derived field is ``None`` (never invented) and\n``reason`` says why. ``state = \"ok\"`` relays the persisted content\nverbatim: population (``SYNTHETIC`` shown as-is), the synthetic spot,\nthe pricing assumptions, the per-(expiration, trading_class) groups and\nthe displayed row budget.",
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "assumptions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assumptions"
          },
          "coverage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "engine_version": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "expirations": {
            "items": {
              "$ref": "#/components/schemas/OptionChainExpiration"
            },
            "title": "Expirations",
            "type": "array"
          },
          "population": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population"
          },
          "reason": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "row_budget": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Budget"
          },
          "snapshot_version": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Version"
          },
          "spot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot"
          },
          "state": {
            "enum": [
              "ok",
              "empty"
            ],
            "title": "State",
            "type": "string"
          },
          "underlying": {
            "minLength": 1,
            "title": "Underlying",
            "type": "string"
          },
          "value_nature": {
            "anyOf": [
              {
                "const": "THEORETICAL",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Nature"
          }
        },
        "required": [
          "state",
          "snapshot_version",
          "as_of",
          "population",
          "underlying",
          "engine_version",
          "value_nature",
          "spot",
          "assumptions",
          "expirations",
          "row_budget",
          "coverage",
          "reason"
        ],
        "title": "OptionChainResponse",
        "type": "object"
      },
...
      "SimulationPreviewRequest": {
        "additionalProperties": false,
        "description": "One declared structure plus its assumptions.\n\nLegs ARE the engine's own :class:`OptionLeg` contract (strict: signed\nnon-zero quantity, strike required on CALL/PUT and forbidden on STOCK,\nnon-negative decimal premium, positive integer multiplier) — the wire\nnever redefines the calculation contract.",
        "properties": {
          "assumptions": {
            "$ref": "#/components/schemas/SimulationAssumptions"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/OptionLeg"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Legs",
            "type": "array"
          }
        },
        "required": [
          "legs",
          "assumptions"
        ],
        "title": "SimulationPreviewRequest",
        "type": "object"
      },
      "SimulationPreviewResponse": {
        "additionalProperties": false,
        "description": "The full preview result (nothing persisted, nothing transactional).",
        "properties": {
          "assumptions": {
            "additionalProperties": true,
            "title": "Assumptions",
            "type": "object"
          },
          "breakevens": {
            "items": {
              "$ref": "#/components/schemas/SimulationBreakeven"
            },
            "title": "Breakevens",
            "type": "array"
          },
          "calculations": {
            "additionalProperties": true,
            "title": "Calculations",
            "type": "object"
          },
          "defined_risk": {
            "additionalProperties": true,
            "title": "Defined Risk",
            "type": "object"
          },
          "max_gain_on_grid": {
            "$ref": "#/components/schemas/SimulationExtreme"
          },
          "max_loss_on_grid": {
            "$ref": "#/components/schemas/SimulationExtreme"
          },
          "payoff_points": {
            "items": {
              "$ref": "#/components/schemas/SimulationPayoffPoint"
            },
            "title": "Payoff Points",
            "type": "array"
          },
          "scenario_grid": {
            "items": {
              "items": {
                "items": {
                  "minLength": 1,
                  "type": "string"
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "Scenario Grid",
            "type": "array"
          },
          "scenario_spot_grid": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Scenario Spot Grid",
            "type": "array"
          },
          "scenario_time_grid_years": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Scenario Time Grid Years",
            "type": "array"
          },
          "value_nature": {
            "const": "THEORETICAL",
            "title": "Value Nature",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "value_nature",
          "defined_risk",
          "payoff_points",
          "breakevens",
          "max_gain_on_grid",
          "max_loss_on_grid",
          "scenario_spot_grid",
          "scenario_time_grid_years",
          "scenario_grid",
          "calculations",
          "assumptions",
          "warnings"
        ],
        "title": "SimulationPreviewResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local, analysis-only API of Vertex One. It serves the canonical contracts of vertex_core and delegates every verdict to the single AdviceEngine. No transactional capability exists; the human remains the sole decision maker, acting outside Vertex.",
    "title": "Vertex One API",
    "version": "vertex_core@0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analysis/{instrument}": {
      "get": {
        "description": "Serve the LAST ``analysis/{instrument}`` snapshot exactly as persisted.\n\nThe API relays the worker's published dossier — the canonical\n``AdviceResult`` of the single ``AdviceEngine`` with its ten gates and\nreason codes, the validated OHLCV bars, the fusion evidence rail and the\n``THEORETICAL`` scenario block (or its typed absence reason) — and\ncomputes nothing. With no snapshot ever published for this instrument\nthe answer is a 200 with ``state = \"empty\"``: absent stays absent,\nnothing is invented.",
        "operationId": "get_analysis",
        "parameters": [
          {
            "in": "path",
            "name": "instrument",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Instrument",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Last published analysis dossier for one instrument (or honest empty state)"
      }
    },
...
    "/api/v1/options/{underlying}/chain": {
      "get": {
        "description": "Serve the LAST ``option_chain/{underlying}`` snapshot exactly as persisted.\n\nThe API relays the worker's published content — per-(expiration,\ntrading_class) groups with complete contract identities, verbatim quotes\nand their quality, the worker's Vertex IV/Greeks (``THEORETICAL``, with\ntheir ``CalculationRecord`` lineage) or their typed refusal reasons, the\ncoverage account and the displayed row budget — and computes nothing.\nWith no snapshot ever published for this underlying the answer is a 200\nwith ``state = \"empty\"``: absent stays absent, nothing is invented.",
        "operationId": "get_option_chain",
        "parameters": [
          {
            "in": "path",
            "name": "underlying",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Underlying",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionChainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Last published option chain for one underlying (or honest empty state)"
      }
    },
    "/api/v1/simulations/preview": {
      "post": {
        "description": "Run one THEORETICAL simulation preview through vertex_core.\n\nOrchestration only: the mandatory ``defined_risk_check``, the exact\n``payoff_at_expiry``, the authority-certified breakevens and the bounded\n``scenario_grid`` all run inside ``vertex_core.calculations.options`` on\na worker thread (``run_in_threadpool`` — the event loop never computes).\nNothing is persisted; nothing here is, or ever becomes, an order.",
        "operationId": "post_simulations_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          },
          "422": {
            "description": "Rejected fail-closed with the exact machine-readable reason: either a wire-contract violation, or the defined-risk verifier's code (e.g. OUTSIDE_CLOSED_CATALOG, UNCOVERED_SHORT_UPSIDE_TAIL, VERTICAL_DEBIT_NOT_BELOW_WIDTH), or a typed calculation-domain violation from vertex_core."
          }
        },
        "summary": "THEORETICAL preview of one declared structure (no persistence)"
      }
    },
...
  }
}